+++ src/test/java/com/lukgru/galgo/runner/fitness/SimpleFitnessCalculatorTest.java
@@ -9,6 +9,7 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import static org.hamcrest.CoreMatchers.hasItem;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -22,19 +23,20 @@ public class SimpleFitnessCalculatorTest {
     public void shouldCalculateForEachInPopulation() {
         //given
         Population<Integer> population = new Population<>(Arrays.asList(1, 2, 3));
-        FitnessFunction<Integer> fitness = new FitnessFunction<>(i -> (i*i) + (7*i) - 12, 0);
+        FitnessFunction<Integer> fitness = new FitnessFunction<>(i -> (double) (i * i) + (7 * i) - 12, 0.0);
 
         //when
         new SimpleFitnessCalculator<>(fitness).compute(population);
-        
+
         //then
         Collection<Individual<Integer>> individuals = population.getIndividuals();
-        List<Integer> listOfComputedFitnessScores = individuals.stream()
+
+        List<Double> listOfComputedFitnessScores = individuals.stream()
                 .map(Individual::getFitnessScore)
                 .collect(Collectors.toList());
-        assertThat(listOfComputedFitnessScores, hasItem(-4));
-        assertThat(listOfComputedFitnessScores, hasItem(6));
-        assertThat(listOfComputedFitnessScores, hasItem(18));
+        assertThat(listOfComputedFitnessScores, hasItem(-4.0));
+        assertThat(listOfComputedFitnessScores, hasItem(6.0));
+        assertThat(listOfComputedFitnessScores, hasItem(18.0));
     }
 
 }
